+++ CHANGELOG.md
@@ -41,6 +41,7 @@
   - A compendium has been added with 2 macros
     - `Create an Elflines Online Character`: This macro will create a blank Elfline Online Character with skills as defined from the Elflines Online Skill List in the Elfline Online compendium released by RTG.
     - `Create Elflines Online Armory`: This macro will create a folder of items as defined from the Elflines Online Armory in the Elfline Online compendium released by RTG.
+- Introduced a limited view of the mook sheet when a player only has limited permission on the mook
 
 **Changes**
 - Newly created actors and items will automatically have default icons configured.